Rename to clean changelog
