+++ scripts/release/steps/index.js
@@ -1,5 +1,6 @@
 export { default as bumpPrettier } from "./bump-prettier.js";
 export { default as checkGitStatus } from "./check-git-status.js";
+export { default as cleanChangelog } from "./clean-changelog.js";
 export { default as generateBundles } from "./generate-bundles.js";
 export { default as installDependencies } from "./install-dependencies.js";
 export { default as lintFiles } from "./lint-files.js";
@@ -12,4 +13,3 @@ export { default as updateDependentsCount } from "./update-dependents-count.js";
 export { default as updateVersion } from "./update-version.js";
 export { default as validateNewVersion } from "./validate-new-version.js";
 export { default as waitForBotRelease } from "./wait-for-bot-release.js";
-export { default as cleanChangelogUnreleased } from "./clean-changelog-unreleased.js";
